+++ css/privatsphaere.css
@@ -1,0 +1,3 @@
+.friend-list {
+  padding: 2em 1em;
+}
