+++ einstellungen/up.php
@@ -11,16 +11,18 @@
 secure_session_start();
 if(isset($_POST['allow'], $_POST['uid'])){
   if(login_check($mysqli)){
-    if($stmt = $mysqli->prepare("UPDATE person SET visibility = ? WHERE uid = ?;")){
-      $stmt->bind_param('is', $_POST['allowNotifications'], $_POST['uid']);
+    if($stmt = $mysqli->prepare("UPDATE person SET privacySettings = ? WHERE uid = ?;")){
+      $stmt->bind_param('is', $_POST['allow'], $_POST['uid']);
       $stmt->execute();
-      if($stmt->affected_rows == 1){
+      if($stmt->error == ""){
+        // Rebuild user session array
+        $_SESSION['user']['privacySettings'] = getUser($_SESSION['user']['uid'], $mysqli)['privacySettings'];
         success([]);
       } else {
-        error('internalError', 500, $stmt->error);
+        error('internalError', 500, "mysqli statement error: " . $stmt->error);
       }
     } else {
-      error('internalError', 500, $mysqli->error);
+      error('internalError', 500, "general mysqli error: " . $mysqli->error);
     }
   }
 } else {
