+++ freunde/anfragen/_ajax/received-requests.php
@@ -7,21 +7,13 @@
 		if(isset($_GET['amount'])){
 			// fetch a limited amount of data
 			$limit = intval($_GET['amount']);
-			$stmt = $mysqli->prepare('SELECT von FROM anfragen WHERE an = ? LIMIT ?');
-			$stmt->bind_param('si', $_GET['uid'], $limit);
-			$stmt->execute();
-			$stmt->bind_result($an);
-			$stmt->store_result();
+			$stmt = fetch("von", "anfragen", "an", $_GET['uid'], $mysqli, "LIMIT $limit");
 			while($stmt->fetch()){
 				$requests[] = $an;
 			}
 		} else {
 			// fetch all data
-			$stmt = $mysqli->prepare('SELECT von FROM anfragen WHERE an = ?');
-			$stmt->bind_param('s', $_GET['uid']);
-			$stmt->execute();
-			$stmt->bind_result($von);
-			$stmt->store_result();
+			$stmt = fetch("von", "anfragen", "an", $_GET['uid'], $mysqli);
 			while($stmt->fetch()){
 				$requests[] = $von;
 			}
@@ -29,18 +21,20 @@
 		$output = "";
 		if(!empty($requests)) :
 			foreach($requests as $element){
-				$fetched_user = getMinimalUser($element, $mysqli);
-				$output .= '<li class="mdl-list__item">
-											<span class="mdl-list__item-primary-content">
-												<img class="mdl-list__item-icon request__user-avatar" src="/users/' . $fetched_user['uid'] . '/avatar.jpg">
-												' . $fetched_user['name'] . '
-											</span>
-											<span class="mdl-list__item-secondary-action request__label--pending">
-												<button class="mdl-button mdl-js-button mdl-js-ripple-effect mdl-button--icon accept-request" data-requestor="' . $fetched_user['uid'] . '">
-													<i class="material-icons">person_add</i>
-												</button>
-											</span>
-										</li>';
+        if(is_visible($element, $_SESSION['user']['uid'], $mysqli)){
+          $fetched_user = getMinimalUser($element, $mysqli);
+  				$output .= '<li class="mdl-list__item">
+  											<span class="mdl-list__item-primary-content">
+  												<img class="mdl-list__item-icon request__user-avatar" src="/users/data/' . $fetched_user['uid'] . '/avatar.jpg">
+  												' . $fetched_user['name'] . '
+  											</span>
+  											<span class="mdl-list__item-secondary-action request__label--pending">
+  												<button class="mdl-button mdl-js-button mdl-js-ripple-effect mdl-button--icon accept-request" data-requestor="' . $fetched_user['uid'] . '">
+  													<i class="material-icons">person_add</i>
+  												</button>
+  											</span>
+  										</li>';
+        }
 			}
       echo $output;
 		else :
