+++ freunde/index.php
@@ -15,17 +15,9 @@
     <body>
       <div class="layout mdl-layout mdl-js-layout mdl-layout--fixed-drawer">
         <div class="drawer mdl-layout__drawer mdl-color--blue-grey-900 mdl-color-text--blue-grey-50">
-          <header class="drawer-header">
-            <img class="avatar" src="<?php echo '/users/' . $user['uid'] . '/avatar.jpg'?>">
-            <div>
-            <div class="flex-container">
-              <span><?php echo $user['email']; ?></span>
-              <div class="mdl-layout-spacer"></div>
-            </div>
-          </div>
-        </header>
-        <?php _getNav('freunde'); ?>
-      </div>
+          <?php require_once(ABS_PATH . '/_sections/drawer.php'); ?>
+          <?php _getNav('freunde'); ?>
+        </div>
         <main class="mdl-layout__content mdl-color--grey-100" meta-username="<?php echo $user['uid'] ?>" meta-directory="<?php echo $user['uid'] ?>">
           <div class="mdl-card container container--margin-top mdl-color--white mdl-shadow--2dp mdl-card--border">
             <h3>Nach Freunden suchen</h3>
@@ -81,10 +73,10 @@
               <div class="mdl-card container mdl-color--white mdl-shadow--2dp mdl-card--border" id="found">
                 <p class="section__title">Auf der Suche nach <b>"<?php echo $_GET['search']?>"</b> haben wir gefunden:</p>
                 <ul class="mdl-list">
-                  <?php foreach($found as $person){ ?>
+                  <?php foreach(filterUsers($_SESSION['user']['uid'], $found, $mysqli) as $person){ ?>
                   <li class="mdl-list__item">
-                    <a href="../users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
-      								<img src="/users/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
+                    <a href="/users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
+      								<img src="/users/data/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
       								<span class="name mdl-color-text--black"><?= $person['name']?></span>
       							</a>
                   </li>
@@ -95,10 +87,10 @@
               <div class="mdl-card container friend-list__container--bottom mdl-color--white mdl-shadow--2dp" id="rest">
                 <p class="section__title">Weitere Menschen, die deine Freunde sein könnten:</p>
                 <ul class="mdl-list">
-                  <?php foreach($rest as $person){ ?>
+                  <?php foreach(filterUsers($_SESSION['user']['uid'], $rest, $mysqli) as $person){ ?>
                   <li class="mdl-list__item">
-                    <a href="../users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
-  										<img src="/users/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
+                    <a href="/users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
+  										<img src="/users/data/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
   										<span class="name mdl-color-text--black"><?= $person['name']?></span>
   									</a>
                   </li>
@@ -125,10 +117,10 @@
                       </li>
                     <?php
       							} else {
-      								foreach($users as $person){ ?>
+      								foreach(filterUsers($_SESSION['user']['uid'], $users, $mysqli) as $person){ ?>
                         <li class="mdl-list__item">
-                            <a href="../users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
-          										<img src="/users/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
+                            <a href="/users/index.php/<?= $person['uid']?>/" class="mdl-list__item-primary-content">
+          										<img src="/users/data/<?= $person['uid']?>/avatar.jpg" class="mdl-list__item-avatar">
           										<span class="name mdl-color-text--black"><?= $person['name']?></span>
           									</a>
                         </li>
