+++ includes/getUserFile.php
@@ -13,7 +13,7 @@
 
   if(login_check($mysqli)){
     if(!is_admin($_SESSION['user']['uid'], $mysqli)){
-      echo success(["json" => file_get_contents(ABS_PATH.'/users/'.$_SESSION['user']['uid'].'/'.$_SESSION['user']['uid'].'.json')]);
+      echo success(["json" => file_get_contents(ABS_PATH.'/users/data/'.$_SESSION['user']['uid'].'/'.$_SESSION['user']['uid'].'.json')]);
     } else {
       echo error('internalError', 500, 'There is no user data for an admin uid');
     }
